requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.13
faiss-cpu>=1.7.4
openai>=1.10.0
pydantic>=2.5.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
pytest>=7.4.3
pytest-asyncio>=0.21.1
python-dotenv>=1.0.0
PyYAML>=6.0.1
tiktoken>=0.5.2
requests>=2.31.0 